Move nested loop join and scan to non-idx folder
